Refactor code and add 'Parser.java', 'Storage.java', 'TaskList.java', 'Ui.java'
Code was grouped together under 'DogeBot.java'.

Didn't make complete use of OOP principles and would be harder to implement new changes.

Refactored code to extract out closely related code as classes to make use of OOP principles under 'Parser.java', 'Storage.java', 'TaskList.java' and 'Ui.java'.
