+++ src/main/java/Ui.java
@@ -1,0 +1,12 @@
+public class Ui {
+    public void intro() {
+        final String LOGO = "    ___\n"
+            + " __/_  `.  .-\"\"\"-."         + "           |                      |             |   \n"
+            + " \\_,` | \\-'  /   )`-')"      + "    _` |   _ \\    _` |   _ \\  __ \\    _ \\   __| \n"
+            + "  \"\") `\"`    \\  ((`\"`"    + "    (   |  (   |  (   |   __/  |   |  (   |  |   \n"
+            + " ___Y  ,    .'7 /|"            + "      \\__,_| \\___/  \\__, | \\___| _.__/  \\___/  \\__| \n"
+            + "(_,___/...-` (_/_/"            + "                    |___/";
+
+        System.out.println(LOGO + "\nHi ! I'm DogeBot \nHow may I help you today ?\n");
+    }
+}
